requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.7
greenlet==3.4.0
idna==3.11
playwright==1.58.0
pydantic==2.12.5
pydantic-core==2.41.5
pyee==13.0.1
python-dotenv==1.2.2
requests==2.33.1
soupsieve==2.8.3
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
